+++ tt_metal/impl/tensor/spec/tensor_spec.cpp
@@ -213,7 +213,7 @@ TensorSpec TensorSpec::sharded(NdShardSpec nd_shard_spec, ShardShapeAlignment sh
                              ? page_config().get_required_shard_shape_alignment()
                              : page_config().get_recommended_shard_shape_alignment(data_type());
         auto& shard_shape = nd_shard_spec.shard_shape;
-        for (int dim = 1; dim <= alignment.size(); dim++) {
+        for (size_t dim = 1; dim <= alignment.size(); dim++) {
             shard_shape[-dim] = round_up(shard_shape[-dim], alignment[-dim]);
         }
     }
